[EPT-1715_ADD_LIST_LAYOUT_IN_PREP_FOR_MODAL] Rename pagesApi
